requirements.txt
PySide6>=6.8.0
emergentintegrations
python-dotenv==1.0.0
aiohttp>=3.8.0,<4.0.0
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
PyAudio>=0.2.14
numpy>=1.24.0
